1. Make tests run faster by caching result of previous run.
multi_thread.cpp: keep time_test a double (not size_t).
sum_i_inv_time.cpp: only execute once when that achieves time_min.

git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2196 0e562018-b8fb-0310-a933-ecf073c7c197

+++ multi_thread/multi_thread.cpp
@@ -186,7 +186,7 @@ namespace {
 		std::cerr << error_msg << std::endl;
 		exit(1);
 	}
-	size_t arg2double(
+	double arg2double(
 		const char* arg       , 
 		double limit          , 
 		const char* error_msg )
